hid debug button
I commented out the debug button. I am doing this to test out whether this is the latest version of the code in production, ie, will it publish to https://sql-practice-app-worker.charlotteachaze.workers.dev/
?

Still need to determine why it publishes there (if it does) instead of to https://sql-practice-app.pages.dev/ , which is where the cloudflare worker points.